+++ qualtran/bloqs/arithmetic/sorting.py
@@ -77,7 +77,7 @@ def _cmp_symb() -> Comparator:
 
 _COMPARATOR_DOC = BloqDocSpec(
     bloq_cls=Comparator,
-    import_line='from qualtran.bloqs.sorting import Comparator',
+    import_line='from qualtran.bloqs.arithmetic.sorting import Comparator',
     examples=(_cmp_symb,),
 )
 
@@ -133,6 +133,6 @@ def _bitonic_sort() -> BitonicSort:
 
 _BITONIC_SORT_DOC = BloqDocSpec(
     bloq_cls=BitonicSort,
-    import_line='from qualtran.bloqs.sorting import BitonicSort',
+    import_line='from qualtran.bloqs.arithmetic.sorting import BitonicSort',
     examples=(_bitonic_sort,),
 )
