+++ qualtran/serialization/bloq.py
@@ -32,7 +32,8 @@ from qualtran import (
     Signature,
     Soquet,
 )
-from qualtran.bloqs import and_bloq, arithmetic, basic_gates, factoring, sorting, swap_network
+from qualtran.bloqs import and_bloq, arithmetic, basic_gates, factoring, swap_network
+from qualtran.bloqs.arithmetic import sorting
 from qualtran.bloqs.util_bloqs import Allocate, ArbitraryClifford, Free, Join, Split
 from qualtran.cirq_interop import CirqGateAsBloq
 from qualtran.protos import args_pb2, bloq_pb2
